code/UE5_SIM_Motors_Streamlit/requirements.txt
pandas==2.1.1
paho-mqtt==1.6.1
numpy==1.25.2
streamlit==1.32.1